More documentation and refactoring
