+++ src/main/java/com/cvent/couchbase/session/CouchbaseHttpSessionProvider.java
@@ -1,75 +1,84 @@
 package com.cvent.couchbase.session;
 
-import com.cvent.couchbase.session.CouchbaseSessionDataStore.CouchbaseSessionData;
-import org.eclipse.jetty.server.session.SessionData;
-import org.glassfish.hk2.api.Injectee;
-import org.glassfish.hk2.api.InjectionResolver;
-import org.glassfish.hk2.api.ServiceHandle;
+import org.glassfish.hk2.api.Factory;
+import org.glassfish.hk2.api.ServiceLocator;
+import org.glassfish.jersey.server.internal.inject.AbstractContainerRequestValueFactory;
+import org.glassfish.jersey.server.internal.inject.AbstractValueFactoryProvider;
+import org.glassfish.jersey.server.internal.inject.MultivaluedParameterExtractorProvider;
+import org.glassfish.jersey.server.internal.inject.ParamInjectionResolver;
+import org.glassfish.jersey.server.model.Parameter;
 
 import javax.inject.Inject;
+import javax.inject.Singleton;
 import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.core.Context;
 
 /**
- * Provides the HttpSession entity for any resource annotated with @CouchbaseSession annotated method parameter
- * 
+ * Provides the CouchbaseSessionData entity for any resource annotated with @CouchbaseSession annotated method parameter
+ *
  * @author bryan
  */
+@Singleton
+public class CouchbaseHttpSessionProvider extends ParamInjectionResolver<CouchbaseSession> {
 
-public class CouchbaseHttpSessionProvider implements InjectionResolver<CouchbaseSession> {
-
-    private final HttpServletRequest req;
-
-    @Inject
-    public CouchbaseHttpSessionProvider(HttpServletRequest request) {
-        this.req = request;
+    public CouchbaseHttpSessionProvider() {
+        super(CouchbaseSessionFactoryProvider.class);
     }
 
-//    @Override
-//    public Injectable<?> getInjectable(ComponentContext ic, final CouchbaseSession session, Parameter parameter) {
-//        if (parameter.getParameterClass().isAssignableFrom(CouchbaseSessionData.class)) {
-//            return () -> {
-//                final HttpServletRequest req = this.req.get();
-//                if (req != null) {
-//                    WrapperSessionCache.CouchbaseSession couchSession =
-//                            (WrapperSessionCache.CouchbaseSession) req.getSession(session.create());
-//                    CouchbaseSessionData couchbaseSessionData = (CouchbaseSessionData) couchSession.getSessionData();
-//                    if (session.write()) {
-//                        couchbaseSessionData.setWrite(true);
-//                    }
-//
-//                    return couchbaseSessionData;
-//                }
-//                return null;
-//            };
-//        }
-//        return null;
-//    }
+    /**
+     * factory provider
+     *
+     */
+    public static class CouchbaseSessionFactoryProvider extends AbstractValueFactoryProvider {
 
-    @Override
-    public Object resolve(Injectee injectee, ServiceHandle<?> root) {
-        if (injectee.getRequiredType().getClass().isAssignableFrom(SessionData.class) && req != null) {
-                CouchbaseSession session = injectee.getParent().getAnnotation(CouchbaseSession.class);
-                WrapperSessionCache.CouchbaseSession couchSession =
-                        (WrapperSessionCache.CouchbaseSession) req.getSession(session.create());
-                CouchbaseSessionData couchbaseSessionData = (CouchbaseSessionData) couchSession.getSessionData();
-                if (session.write()) {
-                    couchbaseSessionData.setWrite(true);
-                }
-                return couchbaseSessionData;
-        } else {
-            return null;
+        /**
+         * Initialize the provider.
+         *
+         * @param mpep                   multivalued map parameter extractor provider.
+         * @param locator                HK2 service locator.
+         */
+
+        @Inject
+        public CouchbaseSessionFactoryProvider(MultivaluedParameterExtractorProvider mpep,
+                                               ServiceLocator locator) {
+            super(mpep, locator, Parameter.Source.UNKNOWN);
+        }
+
+        @Override
+        public Factory<? extends CouchbaseSessionData> createValueFactory(Parameter parameter) {
+            Class<?> paramType = parameter.getRawType();
+            CouchbaseSession annotation = parameter.getAnnotation(CouchbaseSession.class);
+            if (annotation != null && paramType.isAssignableFrom(CouchbaseSessionData.class)) {
+                return new CouchbaseSessionFactory(annotation);
+            } else {
+                return null;
+            }
+        }
+
+    }
+
+    /**
+     * A factory that extracts the session from the http request
+     */
+    public static class CouchbaseSessionFactory extends AbstractContainerRequestValueFactory<CouchbaseSessionData> {
+        @Context
+        private HttpServletRequest request;
+
+        private final CouchbaseSession annotation;
+
+        public CouchbaseSessionFactory(CouchbaseSession annotation) {
+            this.annotation = annotation;
+        }
+
+        @Override
+        public CouchbaseSessionData provide() {
+            WrapperSessionCache.CouchbaseSession couchSession =
+                    (WrapperSessionCache.CouchbaseSession) request.getSession(annotation.create());
+            final CouchbaseSessionData couchbaseSessionData = (CouchbaseSessionData) couchSession.getSessionData();
+            if (annotation.write()) {
+                couchbaseSessionData.setWrite(true);
+            }
+            return couchbaseSessionData;
         }
     }
-
-    @Override
-    public boolean isConstructorParameterIndicator() {
-        return false;
-    }
-
-    @Override
-    public boolean isMethodParameterIndicator() {
-        return true;
-    }
-
-
 }
